requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic[email]
passlib[bcrypt]
python-jose[cryptography]
requests
pytest
httpx
